# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IUG_FINDER_API",
    "version": "1.0.0",
    "description": "IUG Finder API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/reports": {
      "post": {
        "tags": ["IUG Finder Cretae reports"],
        "summary": "User Cretae reports",
        "description": "reports for new user",
        "operationId": "reports For User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "item_type",
                  "color",
                  "report_date",
                  "report_time",
                  "location",
                  "description"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "موجود او مفقود"
                  },
                  "item_type": {
                    "type": "string",
                    "example": "جوال ايفون ١٦"
                  },
                  "color": {
                    "type": "string",
                    "example": "احمر"
                  },
                  "report_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-04-17"
                  },
                  "report_time": {
                    "type": "string",
                    "format": "time",
                    "example": "11:30:00"
                  },
                  "location": {
                    "type": "string",
                    "example": "شارع B - الطابق الثاني"
                  },
                  "description": {
                    "type": "string",
                    "example": "قرب الكافتيريا"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Report created successfully"
                    },
                    "reportId": {
                      "type": "integer",
                      "example": 8
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      },
      "get": {
        "tags": ["IUG Finder Cretae reports"],
        "summary": "get Reports For User",
        "description": "Get all reports based on user role",
        "operationId": "getReports",
        "responses": {
          "200": {
            "description": "Reports fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Reports fetched successfully"
                    },
                    "length": {
                      "type": "integer",
                      "example": 2
                    },
                    "reports": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer", "example": 1 },
                          "item_type": { "type": "string", "example": "مفتاح" },
                          "color": { "type": "string", "example": "أسود" },
                          "report_date": { "type": "string", "format": "date-time", "example": "2025-04-16T21:00:00.000Z" },
                          "report_time": { "type": "string", "format": "time", "example": "11:30:00" },
                          "location": { "type": "string", "example": "مبنى B - الطابق الثاني" },
                          "description": { "type": "string", "example": "المفتاح ضاع قرب الكافتيريا" },
                          "user_id": { "type": "integer", "example": 3 },
                          "created_at": { "type": "string", "format": "date-time", "example": "2025-04-17T12:12:52.000Z" },
                          "status": { "type": "string", "example": "مفقود" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    }
...
    },"/api/reports/{id}": {
      "delete": {
        "tags": ["IUG Finder Cretae reports"],
        "summary": "Delete a report",
        "description": "Allows an admin to delete any report, and a user to delete their own report only.",
        "operationId": "deleteReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the report to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Report deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied. You can only delete your own reports."
          },
          "404": {
            "description": "Report not found"
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    }
  }
}